Tests: Move integration tests to core blocks module
