+++ core-blocks/test/integration/classic-in.html
