+++ core-blocks/test/integration/classic-out.html
