+++ core-blocks/test/integration/one-image-in.html
