+++ core-blocks/test/integration/one-image-out.html
